--- a/wolfcrypt/src/siphash.c
+++ b/wolfcrypt/src/siphash.c
@@ -21,6 +21,10 @@
 
 #include <wolfssl/wolfcrypt/libwolfssl_sources.h>
 
+#ifdef WC_SIPHASH_NO_ASM
+    #define WOLFSSL_NO_ASM
+#endif
+
 #include <wolfssl/wolfcrypt/siphash.h>
 
 #ifdef NO_INLINE