@douzzer 2025-05-20
retest this please
```
Found unhandled org.jenkinsci.plugins.workflow.support.steps.AgentOfflineException exception:
Unable to create live FilePath for wolf-linux-cloud-node-yniezk; wolf-linux-cloud-node-yniezk was marked offline: Connection was broken
```